+++ inst/app/tools/app/report.R
@@ -173,17 +173,6 @@ scrub <- . %>%
   gsub("&lt;!&ndash;/html_preserve&ndash;&gt;","",.)  ## knitr adds this
 
 ## cleanout widgets not needed outside shiny apps
-# cleanout <- function(x) {
-#   gsub("%>%\\s*render\\(\\.\\)", "", x) %>%
-#   gsub("%>%\\s*render", "", .) %>%
-#   gsub("[^A-Za-z]render\\((.*)\\)", "\\1", .) %>%
-#   gsub("^render\\((.*)\\)", "\\1", .)
-# }
-# "test = render(lklklk)" %>% cleanout
-# "%>% render" %>% cleanout
-# "%>% render(.)" %>% cleanout
-# "render(lklklk)" %>% cleanout
-
 cleanout <- . %>%
   gsub("DiagrammeR::renderDiagrammeR", "", .) %>% ## leave for legacy reasons
   gsub("DT::renderDataTable", "", .) %>%          ## leave for legacy reasons