+++ inst/app/tools/data/manage.R
@@ -114,7 +114,7 @@ loadUserData <- function(fname, uFile, ext,
     }
     robj <- robj[1:nrows,]
   } else if (ext == 'csv') {
-    r_data[[objname]] <- loadcsv(uFile, .csv = .csv, header = header, n_max = n_max, sep = sep, saf = man_str_as_factor) %>%
+    r_data[[objname]] <- loadcsv(uFile, .csv = .csv, header = header, n_max = n_max, sep = sep, dec = dec, saf = man_str_as_factor) %>%
       {if (is.character(.)) upload_error_handler(objname, "### There was an error loading the data") else .} %>%
       {set_colnames(., gsub("^\\s+|\\s+$", "", names(.)))}