+++ NEWS.md
@@ -18,3 +18,4 @@
 - Remove \\r and special characters from strings in r_data and r_state 
 - Fix sorting in _R > Report_ for tables created using _Data > Pivot_ and _Data > Explore_ when column headers contain symbols or spaces (thanks @4kammer)
 - Set `error = TRUE` for rmarkdown for consistency with knitr as used in R > Report
+- Correctly handle decimal indicators when loading csv files in _Data > Manage_